+++ src/main/java/io/github/leanish/sqs/codec/algorithms/checksum/Sha256Digestor.java
@@ -7,9 +7,10 @@ package io.github.leanish.sqs.codec.algorithms.checksum;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.util.Base64;
 
 import com.google.errorprone.annotations.Immutable;
+
+import io.github.leanish.sqs.codec.algorithms.encoding.Base64Codec;
 
 /**
  * SHA-256 digest implementation.
@@ -17,13 +18,11 @@ import com.google.errorprone.annotations.Immutable;
 @Immutable
 public class Sha256Digestor implements Digestor {
 
-    private static final Base64.Encoder BASE64_ENCODER = Base64.getUrlEncoder();
-
     @Override
     public String checksum(byte[] payload) {
         MessageDigest digest = digest();
         byte[] hash = digest.digest(payload);
-        return BASE64_ENCODER.encodeToString(hash);
+        return Base64Codec.instance().encodeToString(hash);
     }
 
     private MessageDigest digest() {